Move most fuzzer code to tools/shell/fuzzer.cpp
