+++ tools/shell/include/shell_fuzzer.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace duckdb_shell {
+
+#ifdef DUCKDB_FUZZER
+int RunFuzzer();
+#endif
+
+} // namespace duckdb_shell
